+++ xServer.D/Feature/Database/Tables/ProfileReservationData.cs
@@ -16,7 +16,7 @@ namespace x42.Feature.Database.Tables
         public string Signature { get; set; }
         public int Status { get; set; }
         public string PriceLockId { get; set; }
-        public long ReservationExpirationBlock { get; set; }
+        public int ReservationExpirationBlock { get; set; }
         public bool Relayed { get; set; }
     }
 }
