+++ xServer.D/Feature/X42Client/X42ClientFeature.cs
@@ -31,7 +31,7 @@ namespace x42.Feature.X42Client
         private readonly X42ClientSettings x42ClientSettings;
 
         public ConnectionStatus Status => x42Client.Status;
-        public ulong BlockTIP => x42Client.BlockTIP;
+        public uint BlockTIP => x42Client.BlockTIP;
 
         public X42ClientFeature(
             ServerNodeBase network,
